feat: add support for adding and deleting block field options
The changes in this commit add functionality to add and delete options for block fields. 

Previously, the block fields did not have the ability to have multiple options. With this update, users can now add and delete options for each block field. 

This change improves the flexibility and customization options for block fields, allowing users to create more dynamic and interactive forms.