GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Update vite config (refs/gitbutler/update-vite-config)
   - Pulumi.dev.yaml
   - .gitignore
   - pnpm-lock.yaml
   - packages/templates/package.json
   - packages/helpers/package.json
   - packages/schemas/package.json
   - packages/frontend/package.json
   - packages/backend/package.json
   - package.json

Your previous branch was: refs/heads/master

The sha for that commit was: ed3889569eb691824296c45a681cb4bc818c7513

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch